fix: harden notification subscription updates
